Compile time select algorithm
